+++ idea/testData/quickfix/foldTryCatch/conditional.kt
@@ -1,5 +1,6 @@
 // "Lift assignment out of 'try' expression" "false"
 // ACTION: Make variable mutable
+// ACTION: Remove redundant assignment
 // ERROR: Val cannot be reassigned
 // WITH_RUNTIME