+++ idea/testData/quickfix/removeRedundantAssignment/notRedundant.kt
@@ -1,0 +1,6 @@
+// "Remove redundant assignment" "false"
+fun foo(): Int {
+    var i = 1
+    <caret>i = 2
+    return i
+}