+++ idea/testData/quickfix/removeRedundantAssignment/simpleRef.kt
@@ -1,0 +1,5 @@
+// "Remove redundant assignment" "true"
+fun test(n: Int) {
+    var i: Int
+    <caret>i = n
+}