+++ idea/testData/quickfix/variables/changeMutability/funParameter.kt
@@ -1,4 +1,5 @@
 // "Make variable mutable" "false"
+// ACTION: Remove redundant assignment
 // ERROR: Val cannot be reassigned
 fun fun1(i: Int) {
     <caret>i = 2